Not need to pass join attributes to association build
